requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
